рефакторинг основной страницы чата
