+++ src/components/side_bar/side_bar.ts
@@ -13,3 +13,5 @@ export default function renderSideBarPreview() {
     });
   return outLine;
 }
+
+
