+++ src/pages/500/index.ts
@@ -1,1 +1,1 @@
-export { default as ErrorPage500 } from "./500";
+export { default as ErrorPage500 } from "./500"
