--- src/pages/main/main.css
@@ -1,17 +1,0 @@
-body {
-  margin: 0px;
-}
-div.hide {
-  display: none;
-}
-/** scroll-bar for Chrome */
-::-weblit-scrollbar {
-  width: 6px !important;
-  height: 6px !important;
-}
-html[dir] ::-webkit-scrollbar-thumb {
-  background-color: var(--text-label-color);
-}
-html[dir] .dark ::-webkit-scrollbar-thumb {
-  background-color: var(--text-label-color);
-}
