--- src/pages/main/main.hbs
@@ -1,8 +1,0 @@
-<div>
-  <div id="messages-list">
-    {{{side_bar}}}
-  </div>
-  <div id="chat-list">
-    {{{main}}}
-  </div>
-</div>
