+++ src/pages/settings/settings.ts
@@ -1,10 +1,10 @@
-import hbs_settings from "./settings.hbs";
-import "./settings.css";
-import image from "../../../static/img/placeholder.png";
-import Block from "../../modules/Block";
-import { UserFields, user as User } from "../../modules/User";
-import Router from "../../modules/Router";
-import { ApiAction } from "../../modules/ApiAction";
+import hbs_settings from "./settings.hbs"
+import "./settings.css"
+import image from "../../../static/img/placeholder.png"
+import Block from "../../modules/Block"
+import { UserFields, user as User } from "../../modules/User"
+import Router from "../../modules/Router"
+import { ApiAction } from "../../modules/ApiAction"
 export default class Settings extends Block {
   constructor() {
     super();
@@ -13,13 +13,13 @@ export default class Settings extends Block {
       this.addEvents();   
     });
 
+    const router = new Router(window);
     const api = new ApiAction();
     const req = api.getUser();
-
+    
     req.then( data => {
-      [User.id, User.login, User.display_name, User.first_name, User.second_name, User.email, User.phone] =
-      [data.id, data.login, data.display_name, data.first_name, data.second_name, data.email, data.phone];
-      const router = new Router(window);
+      [User.id, User.login, User.display_name, User.first_name, User.second_name, User.email, User.phone, User.avatar] =
+      [data.id, data.login, data.display_name, data.first_name, data.second_name, data.email, data.phone, data.avatar];
       router._onRoute(window.location.pathname);
     });
   }
@@ -50,6 +50,7 @@ export default class Settings extends Block {
     const changeBtn = <Element>document.querySelector(".change-data");
     const changeAvatar = <Element>( document.querySelector(".set-avatar img") );
     const changePassword = <Element>( document.querySelector(".change-pswd") );
+    const router = new Router(window);
 
     changeBtn &&
       changeBtn.addEventListener("click", (e) => {
@@ -73,11 +74,17 @@ export default class Settings extends Block {
             email: email.value,
             phone: phone.value,
           } as UserFields;
-          api.updateUserInfo(newData);
+          const req = api.updateUserInfo(newData);
           changeBtn.parentElement?.classList.add("locked");
           changeBtn.textContent = "Изменить данные";
+          req.then( data => {
+            if (data.status == 200) {
+              alert("Данные обновлен");
+            }
+          });
         }
       });
+
     changeAvatar &&
       changeAvatar.addEventListener("click", (e) => {
         e.preventDefault();
@@ -88,7 +95,13 @@ export default class Settings extends Block {
           if (!!file) {
             const formData = new FormData();
             formData.append("avatar", file, file.name);
-            api.loadAvatar(formData);
+            const req = api.loadAvatar(formData);
+            req.then( data => {
+              if (data.status == 200) { 
+                console.log('avatar loaded');
+                router._onRoute(window.location.pathname);
+              }
+            });
           }
         };
       });
@@ -97,7 +110,12 @@ export default class Settings extends Block {
         e.preventDefault();
         const oldPswd = prompt("Введите ТЕКУЩИЙ пароль:");
         const newPswd = prompt("Задайте НОВЫЙ пароль:");
-        api.updatePassword(newPswd!, oldPswd!);
+        const req = api.updatePassword(newPswd!, oldPswd!);
+        req.then( data => {
+          if (data.status == 200) { 
+            alert('Пароль обновлен');
+          }
+        });
       });
   }
 }
